Extended Note for Asset Setup
Because of the way relative pathing is set up for Gems, users may encounter a scenario where they have conflicting assets because they are in the same location with the same name within their respective gems. It is worth noting that by adding other namespaces, folders, etc. it can help make their content less likely to have a conflict with other content.